+++ src/main/java/com/raoulvdberge/refinedstorage/api/storage/StorageDiskType.java
@@ -1,5 +1,8 @@
 package com.raoulvdberge.refinedstorage.api.storage;
 
+/**
+ * The storage disk type.
+ */
 public enum StorageDiskType {
     ITEMS,
     FLUIDS