+++ src/main/java/com/raoulvdberge/refinedstorage/integration/jei/RecipeMakerSolderer.java
@@ -2,13 +2,14 @@ package com.raoulvdberge.refinedstorage.integration.jei;
 
 import com.raoulvdberge.refinedstorage.api.solderer.ISoldererRecipe;
 import com.raoulvdberge.refinedstorage.apiimpl.API;
+import mezz.jei.api.IGuiHelper;
 import net.minecraft.item.ItemStack;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public final class RecipeMakerSolderer {
-    public static List<RecipeWrapperSolderer> getRecipes() {
+    public static List<RecipeWrapperSolderer> getRecipes(IGuiHelper guiHelper) {
         List<RecipeWrapperSolderer> recipes = new ArrayList<>();
 
         for (ISoldererRecipe recipe : API.instance().getSoldererRegistry().getRecipes()) {
@@ -20,7 +21,7 @@ public final class RecipeMakerSolderer {
 
             ItemStack output = recipe.getResult();
 
-            recipes.add(new RecipeWrapperSolderer(inputs, output));
+            recipes.add(new RecipeWrapperSolderer(guiHelper, recipe.getDuration(), inputs, output));
         }
 
         return recipes;