+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fixed Grid crash (raoulvdberge)
 - Fixed Fluid Grid not formatting large quantities correctly (raoulvdberge)
 - Small performance improvement: only sort the storages when needed (raoulvdberge)
+- Display progress bar on JEI recipes for the Solderer (raoulvdberge)
 
 ### 1.4.2
 - Updated Forge to 2261 (raoulvdberge)